+++ ci/pkg/ci/ci.go
@@ -14,6 +14,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strconv"
 	"strings"
 
 	"github.com/fairwindsops/insights-plugins/trivy/pkg/models"
@@ -285,10 +286,10 @@ func SendResults(reports []ReportInfo, resources []Resource, configurationObject
 		return results, err
 	}
 
-	for idx, actionItem := range results.ActionItems {
+	for idx, actionItem := range results.NewActionItems {
 		for _, resource := range resources {
 			if resource.Kind == actionItem.ResourceKind && resource.Name == actionItem.ResourceName {
-				results.ActionItems[idx].Notes = fmt.Sprintf("Resource was found in file: %s with comment of %s", resource.Filename, resource.FileComment)
+				results.NewActionItems[idx].Notes = fmt.Sprintf("Resource was found in file: %s with comment of %s", resource.Filename, resource.FileComment)
 				break
 			}
 		}
@@ -303,7 +304,7 @@ func SaveJUnitFile(results ScanResults, configurationObject Configuration) error
 	if configurationObject.Options.JUnitOutput != "" {
 		cases := make([]formatter.JUnitTestCase, 0)
 
-		for _, actionItem := range results.ActionItems {
+		for _, actionItem := range results.NewActionItems {
 			cases = append(cases, formatter.JUnitTestCase{
 				Name: actionItem.ResourceName + ": " + actionItem.Title,
 				Failure: &formatter.JUnitFailure{
@@ -313,10 +314,16 @@ func SaveJUnitFile(results ScanResults, configurationObject Configuration) error
 			})
 		}
 
+		for _, actionItem := range results.FixedActionItems {
+			cases = append(cases, formatter.JUnitTestCase{
+				Name: actionItem.ResourceName + ": " + actionItem.Title,
+			})
+		}
+
 		testSuites := formatter.JUnitTestSuites{
 			Suites: []formatter.JUnitTestSuite{
 				{
-					Tests:     len(results.ActionItems),
+					Tests:     len(results.NewActionItems) + len(results.FixedActionItems),
 					TestCases: cases,
 				},
 			},
@@ -336,10 +343,23 @@ func SaveJUnitFile(results ScanResults, configurationObject Configuration) error
 	return nil
 }
 
+func getSeverity(severityString string) float64 {
+	if severityString == "danger" {
+		return 0.66
+	}
+	severity, err := strconv.ParseFloat(severityString, 64)
+	if err != nil {
+		panic(err)
+	}
+	return severity
+}
+
 // CheckScore checks if the score meets all of the thresholds.
 func CheckScore(results ScanResults, configurationObject Configuration) error {
-	if configurationObject.Options.ScoreThreshold > results.Score || configurationObject.Options.ScoreChangeThreshold < results.BaselineScore-results.Score {
-		logrus.Infof("Score is out of bounds, please fix some Action Items: %v", results.ActionItems)
+	if len(results.NewActionItems) > configurationObject.Options.NewActionItemThreshold || funk.MaxFloat64(funk.Map(results.NewActionItems, func(ai actionItem) float64 {
+		return ai.Severity
+	}).([]float64)).(float64) >= getSeverity(configurationObject.Options.SeverityThreshold) {
+		logrus.Infof("Score is out of bounds, please fix some Action Items: %v", results.NewActionItems)
 		return errors.New(ScoreOutOfBoundsMessage)
 	}
 
@@ -347,17 +367,42 @@ func CheckScore(results ScanResults, configurationObject Configuration) error {
 }
 
 // ProcessHelmTemplates turns helm into yaml to be processed by Polaris or the other tools.
-func ProcessHelmTemplates(configurationObject Configuration) error {
+func ProcessHelmTemplates(configurationObject Configuration, configFolder string) error {
 	for _, helmObject := range configurationObject.Manifests.Helm {
 		err := util.RunCommand(exec.Command("helm", "dependency", "update", helmObject.Path), "Updating dependencies for "+helmObject.Name)
 		if err != nil {
 			return err
 		}
-		err = util.RunCommand(exec.Command("helm", "template", helmObject.Name, helmObject.Path, "-f", helmObject.VariableFile, "--output-dir", configurationObject.Manifests.FolderName+"/"+helmObject.Name), "Templating: "+helmObject.Name)
+		params := []string{
+			"template", helmObject.Name,
+			helmObject.Path,
+
+			"--output-dir",
+			configFolder + helmObject.Name,
+		}
+		if helmObject.VariableFile != "" {
+			params = append(params, "-f",
+				helmObject.VariableFile)
+		}
+		for variable, value := range helmObject.Variables {
+			params = append(params, "--set", fmt.Sprintf("%s=%s", variable, value))
+		}
+		err = util.RunCommand(exec.Command("helm", params...), "Templating: "+helmObject.Name)
 		if err != nil {
 			return err
 		}
 	}
 	return nil
 
+}
+
+// CopyYaml adds all Yaml found in a given spot into the manifest folder.
+func CopyYaml(configurationObject Configuration, configFolder string) error {
+	for _, path := range configurationObject.Manifests.YamlPaths {
+		err := util.RunCommand(exec.Command("cp", "-r", path, configFolder), "Copying yaml file")
+		if err != nil {
+			return err
+		}
+	}
+	return nil
 }