+++ ci/CHANGELOG.md
@@ -1,5 +1,15 @@
-# 0.1.1
+# Changelog
+
+## 0.2.0
+
+* New config format
+* Send Kubernetes Resources to be saved
+* Base results based on new action items instead of "Score"
+
+## 0.1.1
+
 * Process helm templates
 
-# 0.1.0
+## 0.1.0
+
 * Initial release